updated unit test and sol files for FeeConfiguration for Governance DAO
