+++ contracts/common/ISharedModel.sol
@@ -14,6 +14,7 @@ interface ISharedModel {
         bool isPrivate;
         string description;
         string[] webLinks;
+        FeeConfig feeConfig;
     }
 
     struct MultiSigCreateDAOInputs {
